#236: Cosntant 타입 위치 수정
- 리뷰를 반영하여 정적인 값인 Constant는 EditToDoViewController로 옮겼습니다.
- 관련 코드리뷰 링크: https://github.com/ToDo-Garden/ToDo-Garden/pull/239#discussion_r1648549683
